+++ codes_doc.go
@@ -1,0 +1,3 @@
+package sshserver
+
+//go:generate containerssh-generate-codes